Widget now available in Qt Designer through TimelinePlugin.h. Refactorings for that purpose.

+++ src/timeline/SlideBarItem.cc
@@ -57,7 +57,7 @@ void SlideBarItem::setIndexSliderHeight(int height) {
     slider->setHeight(height);
 }
 
-void SlideBarItem::setOrderedWidth(qreal width) {
+void SlideBarItem::setOrderedWidth(int width) {
     this->ordered_width = width;
     //update();
 }
